rearrange requirements
