--- src/ewoks/_requirements/models/conda.py
@@ -1,13 +1,0 @@
-from typing import Literal
-
-from .base import BaseManagerInfo
-from .base import BaseRequirements
-
-
-class CondaManagerInfo(BaseManagerInfo):
-    name: Literal["conda"] = "conda"
-
-
-class CondaRequirements(BaseRequirements):
-    manager: CondaManagerInfo
-    environment: dict
