--- src/ewoks/_requirements/models/pixi.py
@@ -1,13 +1,0 @@
-from typing import Literal
-
-from .base import BaseManagerInfo
-from .base import BaseRequirements
-
-
-class PixiManagerInfo(BaseManagerInfo):
-    name: Literal["pixi"] = "pixi"
-    lockfile: str
-
-
-class PixiRequirements(BaseRequirements):
-    manager: PixiManagerInfo
