+++ src/ewoks/_requirements/poetry.py
@@ -1,16 +1,30 @@
 import importlib.metadata
 import os
 import sys
+from typing import Any
+from typing import Dict
+from typing import List
+from typing import Literal
 from typing import Optional
 
-from ..metadata.gather import gather_requirements
-from ..models.poetry import PoetryRequirements
-from .utils.base import BaseManager
+from .utils.base_manager import BaseManager
+from .utils.base_manager import BaseManagerInfo
+from .utils.base_manager import BaseRequirements
+
+
+class PoetryManagerInfo(BaseManagerInfo):
+    name: Literal["poetry"] = "pip"
+    requirements: List[str]
+
+
+class PoetryRequirements(BaseRequirements):
+    manager: PoetryManagerInfo
 
 
 class PoetryManager(BaseManager):
     NAME = "poetry"
     PRIORITY = 2
+    REQUIREMENTS_MODEL = PoetryRequirements
 
     def __init__(self, *command: str) -> None:
         if not command:
@@ -28,15 +42,11 @@ class PoetryManager(BaseManager):
         """Manager is explicitly active."""
         return "POETRY_ACTIVE" in os.environ
 
-    def _gather_requirements(self, manager_version: str) -> PoetryRequirements:
+    def _gather_requirements(self) -> Dict[str, Any]:
         output = self._check_output("export", "--without-hashes")
         requirements = output.strip().splitlines()
 
-        return gather_requirements(
-            manager_name=self.NAME,
-            manager_version=manager_version,
-            requirements=requirements,
-        )
+        return {"requirements": requirements}
 
     def _install_requirements(self, requirements: PoetryRequirements) -> None:
         text = "\n".join(requirements.requirements)
