+++ src/ewoks/_requirements/utils/metadata/metadata_models.py
@@ -5,6 +5,21 @@ from pydantic import BaseModel
 from pydantic import Field
 
 
+class SystemInfo(BaseModel):
+    system: str
+    release: str
+    version: str
+    machine: str
+    processor: str
+
+
+class PythonInfo(BaseModel):
+    version: str
+    implementation: str
+    compiler: str
+    build: str
+
+
 class GitInfo(BaseModel):
     commit: str
     remote: Optional[str] = None
